Apply consistent currency formatting across all transaction displays
- Added formatCurrencyDisplay to shared formatters utility
- Updated TransactionDetailDialog to use shared formatter (removed duplicate code)
- Updated TransactionRow to use new formatter and monospace font
- Consistent formatting: [sign]number space currency (e.g., '-1,250.00 ₪')
- Applied monospace font to both transaction list and detail dialog
- Ensures perfect alignment throughout the entire application
- DRY principle: single source of truth for currency formatting
